requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
drf-writable-nested==0.6.3
drf-yasg==1.20.0
executing==0.8.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
phonenumberslite==8.12.43
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
